prefix all the glog flags with `tt_`
If teltech/glog and golang/glog are imported to the same binary,
the conflict and cause a panic. I can't figure out how to stop revel
from including golang/glog in the generated app/tmp/main.go file. So,
I prefixed the flags here to fix the panic.

I'm not sure if this is the proper solution, or the best solution.